fix(DTO)/changes-structure
Minor structural changes that fix the code

+++ BLL/DTOs/StoreInventoryDTO.cs
@@ -5,5 +5,5 @@ public class StoreInventoryDTO
     public required string ProductName;
     public int ProductId;
     public int Quantity;
-    public int Price;
+    public decimal Price;
 }
